requirements.txt
line-bot-sdk
bs4
flask
pymongo
datetime
pandas
SnowNLP
emoji
pyshorteners
scipy
schedule
openpyxl
arrow
APScheduler==3.0.0
